simulacao/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==1.6.1
httpx==0.27.2
aiosqlite==0.20.0
python-multipart==0.0.9
